compnents seperated
