+++ src/components/Theater/Theater.scss
@@ -1,25 +1,3 @@
-.movie-cards-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-.movie-card {
-    border: 1px solid black;
-    margin: 1rem;
-
-    &__poster {
-        width: 150px;
-        height: 200px;
-    }
-
-    &__title {
-        display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    }
-}
-
 .theater-text {
     &__header {
         text-align: center;
